requirements.txt
numpy==1.19.3
matplotlib==3.3.3
Pillow==8.0.1
keras==2.4.3
tensorflow==2.4.0